+++ src/test/java/net/causw/application/user/UserServiceTest.java
@@ -18,24 +18,17 @@ import net.causw.adapter.persistence.user.UserAdmission;
 import net.causw.application.dto.post.PostsResponseDto;
 import net.causw.application.dto.user.UserPostsResponseDto;
 import net.causw.application.dto.user.UserResponseDto;
-import net.causw.application.dto.user.UserUpdateRoleRequestDto;
 import net.causw.application.dto.util.dtoMapper.PostDtoMapper;
 import net.causw.application.dto.util.dtoMapper.UserDtoMapper;
 import net.causw.application.excel.UserExcelService;
 import net.causw.application.pageable.PageableFactory;
-import net.causw.domain.exceptions.ErrorCode;
-import net.causw.domain.exceptions.UnauthorizedException;
 import net.causw.domain.model.enums.user.Role;
 import net.causw.domain.model.enums.user.UserState;
 
-import net.causw.domain.model.util.MessageUtil;
 import net.causw.domain.model.util.ObjectFixtures;
 import net.causw.domain.model.util.StaticValue;
-import net.causw.domain.policy.domain.RolePolicy;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -45,7 +38,6 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyList;
@@ -227,165 +219,5 @@ class UserServiceTest {
 
     }
 
-  }
-
-  @Nested
-  class GrantUserRoleTest {
-
-    private final User grantor = ObjectFixtures.getUser();
-    private final User grantee = ObjectFixtures.getUser();
-    private final String granteeId = "dummyGranteeId";
-
-    private List<User> generateUsersByRole(Role role) {
-      return Stream.of(ObjectFixtures.getUser(), ObjectFixtures.getUser(), ObjectFixtures.getUser())
-              .peek(user -> user.setRoles(Set.of(role))).collect(Collectors.toList());
-    }
-
-    private static Set<Role> getNonGrantableRoles() {
-      return EnumSet.allOf(Role.class).stream()
-              .filter(role -> !RolePolicy.GRANTABLE_ROLES.contains(role))
-              .collect(Collectors.toSet());
-    }
-
-    private static Set<Role> getSpecialRoles() {
-      return EnumSet.allOf(Role.class).stream()
-              .filter(role -> !role.equals(Role.COMMON))
-              .collect(Collectors.toSet());
-    }
-
-    private static Set<Role> getGrantableRolesWithoutEdge() {
-      return RolePolicy.GRANTABLE_ROLES.stream()
-              .filter(role -> !role.equals(Role.PRESIDENT))
-              .collect(Collectors.toSet());
-    }
-
-    private static Set<Role> getGrantableAndUniqueRolesWithoutEdge() {
-      return RolePolicy.GRANTABLE_ROLES.stream()
-              .filter(role -> !role.equals(Role.PRESIDENT))
-              .filter(Role::isUnique)
-              .collect(Collectors.toSet());
-    }
-
-    void assertServiceSuccess(User grantor, Role targetRole) {
-      UserUpdateRoleRequestDto userUpdateRoleRequestDto = new UserUpdateRoleRequestDto(String.valueOf(targetRole));
-      userService.grantUserRole(grantor, granteeId, userUpdateRoleRequestDto);
-      assertThat(grantor.getRoles()).isEqualTo(Set.of(Role.COMMON));
-      assertThat(grantee.getRoles()).isEqualTo(Set.of(targetRole));
-    }
-
-    void assertValidatorFail(User grantor, Role targetRole) {
-      UserUpdateRoleRequestDto userUpdateRoleRequestDto = new UserUpdateRoleRequestDto(String.valueOf(targetRole));
-      assertThatThrownBy(() -> userService.grantUserRole(grantor, granteeId, userUpdateRoleRequestDto))
-              .isInstanceOf(UnauthorizedException.class)
-              .hasMessageContaining(MessageUtil.GRANT_ROLE_NOT_ALLOWED)
-              .extracting("errorCode")
-              .isEqualTo(ErrorCode.GRANT_ROLE_NOT_ALLOWED);
-    }
-
-    @BeforeEach
-    void setUp() {
-      //when
-      when(userRepository.findById(granteeId)).thenReturn(Optional.of(grantee));
-    }
-
-    @ParameterizedTest
-    @MethodSource("getNonGrantableRoles")
-    @DisplayName("위임 권한이 위임 가능 대상이 아닐 경우 실패")
-    void a_Failure(Role role) {
-      // given
-      grantor.setRoles(Set.of(role));
-      grantee.setRoles(Set.of(Role.COMMON));
-
-      // when & then
-      assertValidatorFail(grantee, role);
-    }
-
-    @Test
-    @DisplayName("위임자가 위임 권한을 가지고 있지 않을 경우 실패")
-    void b_Failure() {
-      // given
-      grantor.setRoles(Set.of(Role.COMMON));
-      grantee.setRoles(Set.of(Role.COMMON));
-
-      // when & then
-      assertValidatorFail(grantor, RolePolicy.GRANTABLE_ROLES.iterator().next());
-    }
-
-    @ParameterizedTest
-    @MethodSource("getGrantableRolesWithoutEdge")
-    @DisplayName("피위임자가 특수 권한일 경우 실패(특수 조건을 가진 권한 제외)")
-    void c_Failure2(Role role) {
-      for (Role specialRole : getSpecialRoles()) {
-        // given
-        grantor.setRoles(Set.of(role));
-        grantee.setRoles(Set.of(specialRole));
-
-        // when & then
-        assertValidatorFail(grantor, role);
-      }
-    }
-
-    @ParameterizedTest
-    @MethodSource("getGrantableRolesWithoutEdge")
-    @DisplayName("피위임자가 일반 권한일 경우 성공(특수 조건을 가진 권한 제외)")
-    void a_Success1(Role role) {
-      // given
-      grantor.setRoles(Set.of(role));
-      grantee.setRoles(Set.of(Role.COMMON));
-
-      // when & then
-      assertServiceSuccess(grantor, role);
-    }
-
-    @ParameterizedTest
-    @MethodSource("getGrantableAndUniqueRolesWithoutEdge")
-    @DisplayName("고유 권한 위임 후 해당 권한을 가진 사용자가 없을 경우 성공(특수 조건을 가진 권한 제외)")
-    void a_Success2(Role role) {
-      // given
-      grantor.setRoles(Set.of(role));
-      grantee.setRoles(Set.of(Role.COMMON));
-      List<User> users = generateUsersByRole(role);
-      System.out.println( role + "/" + users.get(0).getRoles());
-      // when
-      when(userRepository.findByRoleAndState(role, UserState.ACTIVE)).thenReturn(users);
-      userService.grantUserRole(grantor, granteeId, new UserUpdateRoleRequestDto(String.valueOf(role)));
-
-      // then
-      users.forEach(user -> assertThat(user.getRoles()).isEqualTo(Set.of(Role.COMMON)));
-    }
-
-    @Test
-    @DisplayName("위임자가 학생회장일 때 피위임자가 부학생회장과 학생회 또는 일반 권한일 경우 성공")
-    void b_Success1() {
-      for (Role role : RolePolicy.ROLES_GRANTABLE_BY_PRESIDENT) {
-        // given
-        grantor.setRoles(Set.of(Role.PRESIDENT));
-        grantee.setRoles(Set.of(role));
-
-        // when & then
-        assertServiceSuccess(grantor, Role.PRESIDENT);
-      }
-    }
-
-    @Test
-    @DisplayName("위임자가 학생회장일 때 학생회장과 부학생회장 그리고 학생회 권한을 가진 사용자가 없을 경우 성공")
-    void b_Success2() {
-      // given
-      grantor.setRoles(Set.of(Role.PRESIDENT));
-      grantee.setRoles(Set.of(Role.COMMON));
-      List<User> presidentUsers = generateUsersByRole(Role.PRESIDENT);
-      List<User> vicePresidentUsers = generateUsersByRole(Role.VICE_PRESIDENT);
-      List<User> councilUsers = generateUsersByRole(Role.COUNCIL);
-
-      // when
-      when(userRepository.findByRoleAndState(Role.PRESIDENT, UserState.ACTIVE)).thenReturn(presidentUsers);
-      when(userRepository.findByRoleAndState(Role.VICE_PRESIDENT, UserState.ACTIVE)).thenReturn(vicePresidentUsers);
-      when(userRepository.findByRoleAndState(Role.COUNCIL, UserState.ACTIVE)).thenReturn(councilUsers);
-      userService.grantUserRole(grantor, granteeId, new UserUpdateRoleRequestDto(String.valueOf(Role.PRESIDENT)));
-
-      // then
-      Stream.of(presidentUsers, vicePresidentUsers, councilUsers).flatMap(Collection::stream)
-              .forEach(user -> assertThat(user.getRoles()).isEqualTo(Set.of(Role.COMMON)));
-    }
   }
 }
